Moved the entire source code of DaphneLib into a subdir "daphne".
- To simplify the merge with the latest state of the main branch.
- Changed the Python import statements accordingly.
