+++ src/api/python/daphne/operator/nodes/do_while_loop.py
@@ -21,28 +21,28 @@
 #
 # -------------------------------------------------------------
 
-from api.python.operator.operation_node import OperationNode
-from api.python.operator.nodes.matrix import Matrix
-from api.python.operator.nodes.scalar import Scalar
-from api.python.operator.nodes.frame import Frame
-from api.python.script_building.dag import OutputType
-from api.python.utils.consts import VALID_INPUT_TYPES
-from api.python.script_building.nested_script import NestedDaphneDSLScript
-from api.python.utils import analyzer
+from daphne.operator.operation_node import OperationNode
+from daphne.operator.nodes.matrix import Matrix
+from daphne.operator.nodes.scalar import Scalar
+from daphne.operator.nodes.frame import Frame
+from daphne.script_building.dag import OutputType
+from daphne.utils.consts import VALID_INPUT_TYPES, VALID_ARITHMETIC_TYPES
+from daphne.script_building.nested_script import NestedDaphneDSLScript
+from daphne.utils import analyzer
 
 from typing import TYPE_CHECKING, Dict, Iterable, Sequence, Tuple, Callable
 import textwrap
+from copy import copy
 
 if TYPE_CHECKING:
     # to avoid cyclic dependencies during runtime
     from context.daphne_context import DaphneContext
 
-
-class WhileLoop(OperationNode):
+class DoWhileLoop(OperationNode):
     def __init__(self, daphne_context: 'DaphneContext', cond: Callable, callback: Callable,
-                 unnamed_input_nodes: Iterable[VALID_INPUT_TYPES] = None) -> 'WhileLoop':
+                 unnamed_input_nodes: Iterable[VALID_INPUT_TYPES] = None) -> 'DoWhileLoop':
         """
-        Operational node that represents while-loop functionality.
+        Operational node that represents do-while-loop functionality.
         Its reserved variable is left unused and is not added ot the generated script.
 
         :param daphne_context:
@@ -51,7 +51,7 @@ class WhileLoop(OperationNode):
         :param unnamed_input_nodes: operation nodes that are up for manipulation
         """
         self.nested_level = 0  # default value
-        # cast the iterable to list for consistency and to avoid addiotnal coping
+        # cast the iterable to list for consistency and to avoid addiotnal copying
         _unnamed_input_nodes = list(unnamed_input_nodes)
         # analyze if the passed functions fulfill the requirements
         if analyzer.get_number_argument(cond) != analyzer.get_number_argument(callback):
@@ -73,6 +73,7 @@ class WhileLoop(OperationNode):
         for node in outer_vars_callback.values():
             if node:
                 _unnamed_input_nodes.append(node)
+
         # TODO: decide if here is the best place for this piece of code: maybe just after the first analysis
         # initiate the output operation nodes
         self._outputs = list()
@@ -89,16 +90,16 @@ class WhileLoop(OperationNode):
             new_matrix_node._source_node = self
             self._outputs.append(new_matrix_node)
 
-        super().__init__(daphne_context, 'while_loop', unnamed_input_nodes=_unnamed_input_nodes,
+        super().__init__(daphne_context, 'do_while_loop', unnamed_input_nodes=_unnamed_input_nodes,
                          output_type=OutputType.NONE)
 
     def __getitem__(self, index) -> Tuple['Matrix']:
-        return self._outputs[index] 
+        return self._outputs[index]  
 
     def code_line(self, var_name: str, unnamed_input_vars: Sequence[str],
                   named_input_vars: Dict[str, str]) -> str:
         """
-        Generates the DaphneDSL code block for while-loop statement.
+        Generates the DaphneDSL code block for do-while-loop statement.
         Here the 'callback' and 'cond' are being
         evaluated and then the code lines for their according functionalities
         are generated and added inside the loop code structure.
@@ -111,7 +112,7 @@ class WhileLoop(OperationNode):
         # handle loop body evaluation and code generation
         callback_outputs = self.callback()
         if not isinstance(callback_outputs, tuple):
-            callback_outputs = (callback_outputs,)
+            callback_outputs = (callback_outputs, )
         # TODO: check why here is used sef.nested_level without +1
         callback_script = NestedDaphneDSLScript(self.daphne_context, self.nested_level)
         callback_names = callback_script.build_code(callback_outputs)
@@ -122,7 +123,7 @@ class WhileLoop(OperationNode):
         # handle loop condition evaluation and code generation
         cond_outputs = self.cond()
         if not isinstance(cond_outputs, tuple):
-            cond_outputs = (cond_outputs,)
+            cond_outputs = (cond_outputs, )
         cond_script = NestedDaphneDSLScript(self.daphne_context, self.nested_level + 1, 'C')
         cond_name = cond_script.build_code(cond_outputs)[0]
         cond_body = cond_script.daphnedsl_script
@@ -130,10 +131,10 @@ class WhileLoop(OperationNode):
         # pack all code lines in the while-loop structure
         multiline_str = str()
         multiline_str += cond_body
-        multiline_str += f"while ({cond_name}) {{\n"
+        multiline_str += "do {\n"
         multiline_str += textwrap.indent(callback_body, prefix="    ")
         multiline_str += textwrap.indent(cond_body, prefix="    ")
-        multiline_str += "}"
+        multiline_str += f"}} while ({cond_name});"
 
         return multiline_str
