+++ src/api/python/daphne/operator/nodes/function.py
@@ -23,12 +23,12 @@
 
 __all__ = ["Function"]
 
-from api.python.operator.operation_node import OperationNode
-from api.python.operator.nodes.matrix import Matrix
-from api.python.script_building.dag import OutputType
-from api.python.script_building.nested_script import NestedDaphneDSLScript
-from api.python.utils import analyzer
-from api.python.utils.consts import VALID_INPUT_TYPES, VALID_ARITHMETIC_TYPES, BINARY_OPERATIONS, TMP_PATH
+from daphne.operator.operation_node import OperationNode
+from daphne.operator.nodes.matrix import Matrix
+from daphne.script_building.dag import OutputType
+from daphne.script_building.nested_script import NestedDaphneDSLScript
+from daphne.utils import analyzer
+from daphne.utils.consts import VALID_INPUT_TYPES, VALID_ARITHMETIC_TYPES, BINARY_OPERATIONS, TMP_PATH
 
 import numpy as np
 from copy import copy
