+++ src/api/python/daphne/utils/consts.py
@@ -25,10 +25,10 @@ import os
 from typing import Union, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from api.python.script_building.dag import DAGNode
-    from api.python.operator.nodes.matrix import Matrix
-    from api.python.operator.nodes.frame import Frame
-    from api.python.operator.nodes.scalar import Scalar
+    from daphne.script_building.dag import DAGNode
+    from daphne.operator.nodes.matrix import Matrix
+    from daphne.operator.nodes.frame import Frame
+    from daphne.operator.nodes.scalar import Scalar
 
 VALID_INPUT_TYPES = Union['DAGNode', str, int, float, bool]
 # These are the operator symbols used in DaphneDSL (not in Python).
